pyproject.toml
[tool.poetry]
name = "vidly-api-django"
version = "0.1.0"
description = "Video Streaming site"
authors = ["HelloSolo <davidalimi148@gmail.com>"]
readme = "README.md"
packages = [{include = "vidly_api_django"}]

[tool.poetry.dependencies]
python = "^3.9"
amqp = "5.1.1"
asgiref = "3.6.0"
async-timeout = "4.0.2"
attrs = "22.2.0"
autopep8 = "2.0.1"
billiard = "3.6.4.0"
black = "22.10.0"
brotli = "1.0.9"
cachecontrol = "0.12.12"
celery = "5.2.7"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
cleo = "2.0.1"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
colorama = "0.4.6"
configargparse = "1.5.3"
coreapi = "2.3.3"
coreschema = "0.0.4"
crashtest = "0.4.1"
cryptography = "39.0.0"
defusedxml = "0.7.1"
distlib = "0.3.6"
dj-database-url = "1.2.0"
django = "4.1.5"
django-cors-headers = "3.13.0"
django-filter = "22.1"
django-storages = "1.13.2"
django-templated-mail = "1.1.1"
django-debug-toolbar = "3.8.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "4.8.0"
djoser = "2.1.0"
docopt = "0.6.2"
drf-nested-routers = "0.93.4"
dulwich = "0.20.50"
exceptiongroup = "1.1.0"
filelock = "3.9.0"
flask = "2.2.2"
flask-basicauth = "0.2.0"
flask-cors = "3.0.10"
gevent = "22.10.2"
geventhttpclient = "2.0.8"
greenlet = "2.0.1"
gunicorn = "20.1.0"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "4.13.0"
iniconfig = "1.1.1"
isodate = "0.6.1"
itsdangerous = "2.1.2"
itypes = "1.2.0"
jaraco-classes = "3.2.3"
jeepney = "0.8.0"
jinja2 = "3.1.2"
jsonschema = "4.17.3"
keyring = "23.13.1"
kombu = "5.2.4"
lockfile = "0.12.2"
locust = "2.14.0"
markupsafe = "2.1.1"
model-bakery = "1.9.0"
more-itertools = "9.0.0"
msgpack = "1.0.4"
msrest = "0.7.1"
mypy-extensions = "0.4.3"
mysql-connector-python = "8.0.31"
mysqlclient = "2.1.0"
oauthlib = "3.2.2"
packaging = "22.0"
pathspec = "0.10.2"
pexpect = "4.8.0"
pillow = "9.4.0"
pipreqs = "0.4.11"
pkginfo = "1.9.3"
platformdirs = "2.6.2"
pluggy = "1.0.0"
poetry = "1.3.1"
poetry-core = "1.4.0"
poetry-plugin-export = "1.2.0"
prompt-toolkit = "3.0.36"
protobuf = "3.20.1"
psutil = "5.9.4"
psycopg2 = "2.9.5"
psycopg2-binary = "2.9.5"
ptyprocess = "0.7.0"
pycodestyle = "2.10.0"
pycparser = "2.21"
pyjwt = "2.6.0"
pyparsing = "3.0.9"
pyrsistent = "0.19.3"
pytest = "7.2.0"
pytest-django = "4.5.2"
pytest-watch = "4.2.0"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
python3-openid = "3.2.0"
pytz = "2022.7"
pyzmq = "24.0.1"
rapidfuzz = "2.13.7"
redis = "4.4.0"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
requests-toolbelt = "0.10.1"
roundrobin = "0.0.4"
secretstorage = "3.3.3"
shellingham = "1.5.0.post1"
six = "1.16.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.3.0"
sqlparse = "0.4.3"
tomli = "2.0.1"
tomlkit = "0.11.6"
trove-classifiers = "2022.12.22"
typing-extensions = "4.4.0"
uritemplate = "4.1.1"
urllib3 = "1.26.13"
vine = "5.0.0"
watchdog = "2.2.0"
wcwidth = "0.2.5"
webencodings = "0.5.1"
werkzeug = "2.2.2"
whitenoise = "6.3.0"
yarg = "0.1.9"
zipp = "3.11.0"
zope-event = "4.6"
zope-interface = "5.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
